Add asset bundles for folders in sideloader mods that do not match an existing one
If .png files are in a folder and the folder name does not match an existing .unity3d file the textures will never load. By making a list of all folders in sideloader mods that contain .png files and then creating a new asset bundle when needed, such textures will load correctly.