[ColorCorrector] Don't override bloom strength by default (#180)
This makes the plugin effectively no-op with the default settings.

Note that this is a breaking change for users who do use the override. I briefly looked for a way to keep the override enabled when the value has been changed from the default, but I couldn't find a nice way. I hope this isn't too much of a problem.

Fixes #179